requirements.txt
anthropic>=0.40.0
openai>=1.0.0
google-genai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0

# For web dashboard
flask>=3.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20
jinja2>=3.1.0

# For enhanced features (optional)
requests>=2.31.0
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
python-docx>=1.0.0
lxml>=4.9.0
